[Windows] Optimize the containerized OVS installation (#6383)
1. Add logic to check the installed OVSext drivers, if only the desired version
    of driver is already installed, skip the installation; otherwise, remove the
    existing drivers and re-install.
2. Add logic to check the installed VC redistributable files, if the existing
    installed vc_redist version is greater or equal than the min required version,
    skip the installation; otherwise re-install with the provided files.
3. Improve the logic updating the env paths by removing duplicated paths.
4. Improve Uninstall-OVS script by removing OVS bin paths from the system path
    after they are deleted.
5. Fix certificate import for a signed OVS driver when Node is running Windows
    Server 2022.

Note that running `.\Install-OVS.ps1` manually is no longer required when using
the containerized OVS installation method, even when using the test-signed
OVS driver that we provide.

+++ hack/windows/Uninstall-OVS.ps1
@@ -5,15 +5,16 @@ Param(
 )
 
 $ErrorActionPreference = "Continue"
-$OVSBinDir = "$OVSInstallDir\usr\bin"
-$OVSDriverDir = "$OVSInstallDir\driver\"
-
+$usrBinPath="$OVSInstallDir\usr\bin"
 if ($DeleteBridges) {
-    $brList = ovs-vsctl.exe list-br
-    foreach ($br in $brList) {
-        Write-Host "Delete OVS Bridge: %s $br"
-        $cmd = "$OVSBinDir\ovs-vsctl.exe --no-wait del-br $br"
-        Invoke-Expression $cmd
+    $ovsDbSock="$OVSInstallDir\var\run\openvswitch\db.sock"
+    if ((Test-Path $usrBinPath) -and (Test-Path $ovsDbSock)){
+        $env:Path="$env:Path;$usrBinPath"
+        $brList = ovs-vsctl.exe list-br
+        foreach ($br in $brList) {
+            Write-Host "Delete OVS Bridge: %s $br"
+            ovs-vsctl.exe --no-wait del-br $br
+        }
     }
 }
 
@@ -37,19 +38,33 @@ if (Get-Service ovsdb-server -ErrorAction SilentlyContinue) {
     }
 }
 # Uninstall OVS kernel driver
-cmd /c "cd $OVSDriverDir && uninstall.cmd"
+$ovsInstalled = $(netcfg -q ovsext) -like "*is installed*"
+if ($ovsInstalled) {
+    netcfg -u ovsext
+}
 if (!$?) {
     Write-Host "Failed to uninstall OVS kernel driver."
     exit 1
 }
 
 # Remove OVS installation dir
-if ($RemoveDir) {
+if ($RemoveDir -and (Test-Path $OVSInstallDir)) {
     Remove-Item -Recurse $OVSInstallDir
     if (!$?) {
         Write-Host "Failed to remove OVS dir: $OVSInstallDir."
         exit 1
     }
 }
+
+# Remove OVS bin paths from environment variables.
+$envPaths = @()
+$usrSbinPath="$OVSInstallDir\usr\sbin"
+foreach ($item in $($env:Path -split ";" | Select-Object -Unique)) {
+    if (($item -ne "$usrBinPath") -and ($item -ne "$usrSbinPath")) {
+        $envPaths += $item
+    }
+}
+$env:Path =$envPaths -join ";"
+[Environment]::SetEnvironmentVariable("Path", $env:Path, [EnvironmentVariableTarget]::Machine)
 
 Write-Host "Uninstall OVS success."
